Make `next_hour` a date_util
