+++ tests/helpers.py
@@ -1,4 +1,3 @@
-import datetime
 from collections.abc import Callable
 from enum import Enum, auto
 from functools import wraps
@@ -8,13 +7,10 @@ from unittest import mock
 from django.contrib.auth import get_user_model
 from graphene_django_extensions.testing import GraphQLClient as BaseGraphQLClient
 
-from common.date_utils import local_datetime
-
 __all__ = [
     "GraphQLClient",
     "ResponseMock",
     "UserType",
-    "next_hour",
 ]
 
 
@@ -115,26 +111,3 @@ class patch_method:
 
     def __exit__(self, *exc_info: object) -> Any:
         return self.patch.__exit__(*exc_info)
-
-
-def next_hour(plus_hours: int = 0, *, plus_minutes: int = 0, plus_days: int = 0) -> datetime.datetime:
-    """
-    Return a timestamp for the next hour.
-
-    Without any arguments, the timestamp will be for the next full hour, any additional arguments will be added to that.
-    e.g.
-    >>> datetime.datetime.now()
-    2021-01-01 12:30:00
-
-    >>> next_hour()
-    13:00
-
-    >>> next_hour(plus_hours=1, plus_minutes=30)
-    14:30
-
-    >>> next_hour(-1)
-    12:00
-    """
-    now = local_datetime()
-    start_of_hour = now.replace(minute=0, second=0, microsecond=0)
-    return start_of_hour + datetime.timedelta(hours=1 + plus_hours, minutes=plus_minutes, days=plus_days)
