+++ tests/test_graphql_api/test_reservation/helpers.py
@@ -5,12 +5,13 @@ from typing import Any
 
 from graphene_django_extensions.testing import build_mutation, build_query
 
+from common.date_utils import next_hour
 from reservation_units.models import ReservationUnit
 from reservations.choices import ReservationTypeChoice
 from reservations.models import Reservation
 from tests.factories import ReservationCancelReasonFactory, ReservationDenyReasonFactory
 from tests.factories.helsinki_profile import MyProfileDataFactory
-from tests.helpers import ResponseMock, next_hour, patch_method
+from tests.helpers import ResponseMock, patch_method
 from users.helauth.clients import HelsinkiProfileClient
 
 reservation_query = partial(build_query, "reservation")
@@ -82,7 +83,7 @@ def get_staff_create_data(
     **overrides: Any,
 ) -> dict[str, Any]:
     if begin is None:
-        begin = next_hour(1)
+        begin = next_hour(plus_hours=1)
     if end is None:
         end = begin + datetime.timedelta(hours=1)
 
@@ -102,7 +103,7 @@ def get_create_data(
     **overrides: Any,
 ) -> dict[str, Any]:
     if begin is None:
-        begin = next_hour(1)
+        begin = next_hour(plus_hours=1)
     if end is None:
         end = begin + datetime.timedelta(hours=1)
 
@@ -157,7 +158,7 @@ def get_staff_adjust_data(
     **overrides: Any,
 ) -> dict[str, Any]:
     if begin is None:
-        begin = next_hour(1)
+        begin = next_hour(plus_hours=1)
     if end is None:
         end = begin + datetime.timedelta(hours=1)
